refactor: separate CLI and MCP server entry points
Split the monolithic index.ts into focused modules to cleanly separate
CLI tooling (skills installer) from MCP server concerns.
